requirements.txt
mutagen
Pillow
pydub
